Move common folder to each layer
[DESC] 각 Layer로 common 폴더에 있었던 파일 이동
