--- app/src/main/java/com/promise09th/mvvmproject/common/ViewType.java
@@ -1,6 +1,0 @@
-package com.promise09th.mvvmproject.common;
-
-public enum ViewType {
-    SEARCH,
-    LOCKER
-}
