+++ app/src/main/java/com/promise09th/mvvmproject/presentation/common/ViewType.java
@@ -1,0 +1,6 @@
+package com.promise09th.mvvmproject.presentation.common;
+
+public enum ViewType {
+    SEARCH,
+    LOCKER
+}
