+++ app/src/main/res/layout/fragment_my_locker.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.promise09th.mvvmproject.presentation.main.ThumbnailViewModel" />
+            type="com.promise09th.mvvmproject.presentation.main.viewmodel.ThumbnailViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
