+++ app/src/main/res/layout/fragment_search_result.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.promise09th.mvvmproject.presentation.main.ThumbnailViewModel" />
+            type="com.promise09th.mvvmproject.presentation.main.viewmodel.ThumbnailViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
