+++ app/src/main/res/layout/item_image_view.xml
@@ -3,13 +3,13 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
-        <import type="com.promise09th.mvvmproject.utils.Utils" />
+        <import type="com.promise09th.mvvmproject.presentation.common.Utils" />
         <variable
             name="viewModel"
-            type="com.promise09th.mvvmproject.presentation.main.ThumbnailViewModel" />
+            type="com.promise09th.mvvmproject.presentation.main.viewmodel.ThumbnailViewModel" />
         <variable
             name="type"
-            type="com.promise09th.mvvmproject.common.ViewType" />
+            type="com.promise09th.mvvmproject.presentation.common.ViewType" />
         <variable
             name="thumbnail"
             type="com.promise09th.mvvmproject.presentation.model.Thumbnail" />
